Organize test_helper a bit
I like to try preventing it from converting into a drawer of misc
unrelated configurations.
