+++ test/support/controller.rb
@@ -1,0 +1,8 @@
+#
+# Base class for controller testing
+#
+class ActionController::TestCase
+  include Devise::TestHelpers
+
+  after { Rails.cache.clear }
+end
